requirements.txt
Django==5.1.6
gunicorn==21.2.0
whitenoise==6.8.2
dj-database-url==2.3.0
psycopg2-binary==2.9.10
python-decouple==3.8
Pillow==11.1.0
